+++ eam/service/service-datacenter/src/main/java/com/dt/platform/datacenter/controller/LayerController.java
@@ -2,6 +2,7 @@ package com.dt.platform.datacenter.controller;
 
 
 import java.util.List;
+import java.util.ArrayList;
 
 import org.springframework.web.bind.annotation.RestController;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -49,7 +50,7 @@ import com.github.foxnic.api.validate.annotations.NotNull;
  * 层级 接口控制器
  * </p>
  * @author 金杰 , maillank@qq.com
- * @since 2022-05-07 21:49:53
+ * @since 2022-07-22 21:41:05
 */
 
 @Api(tags = "层级")
@@ -68,7 +69,7 @@ public class LayerController extends SuperController {
 	@ApiImplicitParams({
 		@ApiImplicitParam(name = LayerVOMeta.ID , value = "主键" , required = true , dataTypeClass=String.class , example = "565457784187916288"),
 		@ApiImplicitParam(name = LayerVOMeta.NAME , value = "名称" , required = false , dataTypeClass=String.class , example = "楼层1"),
-		@ApiImplicitParam(name = LayerVOMeta.AREA_ID , value = "区域" , required = false , dataTypeClass=String.class),
+		@ApiImplicitParam(name = LayerVOMeta.AREA_ID , value = "区域" , required = false , dataTypeClass=String.class , example = "575433967289761792"),
 		@ApiImplicitParam(name = LayerVOMeta.SORT , value = "排序" , required = false , dataTypeClass=Integer.class),
 		@ApiImplicitParam(name = LayerVOMeta.NOTES , value = "备注" , required = false , dataTypeClass=String.class),
 	})
@@ -94,6 +95,17 @@ public class LayerController extends SuperController {
 	@SentinelResource(value = LayerServiceProxy.DELETE , blockHandlerClass = { SentinelExceptionUtil.class } , blockHandler = SentinelExceptionUtil.HANDLER )
 	@PostMapping(LayerServiceProxy.DELETE)
 	public Result deleteById(String id) {
+		this.validator().asserts(id).require("缺少id值");
+		if(this.validator().failure()) {
+			return this.validator().getFirstResult();
+		}
+		// 引用校验
+		Boolean hasRefer = layerService.hasRefers(id);
+		// 判断是否可以删除
+		this.validator().asserts(hasRefer).mustInList("不允许删除当前记录",false);
+		if(this.validator().failure()) {
+			return this.validator().getFirstResult();
+		}
 		Result result=layerService.deleteByIdLogical(id);
 		return result;
 	}
@@ -112,8 +124,43 @@ public class LayerController extends SuperController {
 	@SentinelResource(value = LayerServiceProxy.DELETE_BY_IDS , blockHandlerClass = { SentinelExceptionUtil.class } , blockHandler = SentinelExceptionUtil.HANDLER )
 	@PostMapping(LayerServiceProxy.DELETE_BY_IDS)
 	public Result deleteByIds(List<String> ids) {
-		Result result=layerService.deleteByIdsLogical(ids);
-		return result;
+
+		// 参数校验
+		this.validator().asserts(ids).require("缺少ids参数");
+		if(this.validator().failure()) {
+			return this.validator().getFirstResult();
+		}
+
+		// 查询引用
+		Map<String, Boolean> hasRefersMap = layerService.hasRefers(ids);
+		// 收集可以删除的ID值
+		List<String> canDeleteIds = new ArrayList<>();
+		for (Map.Entry<String, Boolean> e : hasRefersMap.entrySet()) {
+			if (!e.getValue()) {
+				canDeleteIds.add(e.getKey());
+			}
+		}
+
+		// 执行删除
+		if (canDeleteIds.isEmpty()) {
+			// 如果没有一行可以被删除
+			return ErrorDesc.failure().message("无法删除您选中的数据行");
+		} else if (canDeleteIds.size() == ids.size()) {
+			// 如果全部可以删除
+			Result result=layerService.deleteByIdsLogical(canDeleteIds);
+			return result;
+		} else if (canDeleteIds.size()>0 && canDeleteIds.size() < ids.size()) {
+			// 如果部分行可以删除
+			Result result=layerService.deleteByIdsLogical(canDeleteIds);
+			if (result.failure()) {
+				return result;
+			} else {
+				return ErrorDesc.success().message("已删除 " + canDeleteIds.size() + " 行，但另有 " + (ids.size() - canDeleteIds.size()) + " 行数据无法删除").messageLevel4Confirm();
+			}
+		} else {
+			// 理论上，这个分支不存在
+			return ErrorDesc.success().message("数据删除未处理");
+		}
 	}
 
 	/**
@@ -123,12 +170,11 @@ public class LayerController extends SuperController {
 	@ApiImplicitParams({
 		@ApiImplicitParam(name = LayerVOMeta.ID , value = "主键" , required = true , dataTypeClass=String.class , example = "565457784187916288"),
 		@ApiImplicitParam(name = LayerVOMeta.NAME , value = "名称" , required = false , dataTypeClass=String.class , example = "楼层1"),
-		@ApiImplicitParam(name = LayerVOMeta.AREA_ID , value = "区域" , required = false , dataTypeClass=String.class),
+		@ApiImplicitParam(name = LayerVOMeta.AREA_ID , value = "区域" , required = false , dataTypeClass=String.class , example = "575433967289761792"),
 		@ApiImplicitParam(name = LayerVOMeta.SORT , value = "排序" , required = false , dataTypeClass=Integer.class),
 		@ApiImplicitParam(name = LayerVOMeta.NOTES , value = "备注" , required = false , dataTypeClass=String.class),
 	})
 	@ApiOperationSupport( order=4 , ignoreParameters = { LayerVOMeta.PAGE_INDEX , LayerVOMeta.PAGE_SIZE , LayerVOMeta.SEARCH_FIELD , LayerVOMeta.FUZZY_FIELD , LayerVOMeta.SEARCH_VALUE , LayerVOMeta.DIRTY_FIELDS , LayerVOMeta.SORT_FIELD , LayerVOMeta.SORT_TYPE , LayerVOMeta.IDS } )
-	@NotNull(name = LayerVOMeta.ID)
 	@SentinelResource(value = LayerServiceProxy.UPDATE , blockHandlerClass = { SentinelExceptionUtil.class } , blockHandler = SentinelExceptionUtil.HANDLER )
 	@PostMapping(LayerServiceProxy.UPDATE)
 	public Result update(LayerVO layerVO) {
@@ -144,7 +190,7 @@ public class LayerController extends SuperController {
 	@ApiImplicitParams({
 		@ApiImplicitParam(name = LayerVOMeta.ID , value = "主键" , required = true , dataTypeClass=String.class , example = "565457784187916288"),
 		@ApiImplicitParam(name = LayerVOMeta.NAME , value = "名称" , required = false , dataTypeClass=String.class , example = "楼层1"),
-		@ApiImplicitParam(name = LayerVOMeta.AREA_ID , value = "区域" , required = false , dataTypeClass=String.class),
+		@ApiImplicitParam(name = LayerVOMeta.AREA_ID , value = "区域" , required = false , dataTypeClass=String.class , example = "575433967289761792"),
 		@ApiImplicitParam(name = LayerVOMeta.SORT , value = "排序" , required = false , dataTypeClass=Integer.class),
 		@ApiImplicitParam(name = LayerVOMeta.NOTES , value = "备注" , required = false , dataTypeClass=String.class),
 	})
@@ -195,7 +241,7 @@ public class LayerController extends SuperController {
 	@PostMapping(LayerServiceProxy.GET_BY_IDS)
 	public Result<List<Layer>> getByIds(List<String> ids) {
 		Result<List<Layer>> result=new Result<>();
-		List<Layer> list=layerService.getByIds(ids);
+		List<Layer> list=layerService.queryListByIds(ids);
 		result.success(true).data(list);
 		return result;
 	}
@@ -208,7 +254,7 @@ public class LayerController extends SuperController {
 	@ApiImplicitParams({
 		@ApiImplicitParam(name = LayerVOMeta.ID , value = "主键" , required = true , dataTypeClass=String.class , example = "565457784187916288"),
 		@ApiImplicitParam(name = LayerVOMeta.NAME , value = "名称" , required = false , dataTypeClass=String.class , example = "楼层1"),
-		@ApiImplicitParam(name = LayerVOMeta.AREA_ID , value = "区域" , required = false , dataTypeClass=String.class),
+		@ApiImplicitParam(name = LayerVOMeta.AREA_ID , value = "区域" , required = false , dataTypeClass=String.class , example = "575433967289761792"),
 		@ApiImplicitParam(name = LayerVOMeta.SORT , value = "排序" , required = false , dataTypeClass=Integer.class),
 		@ApiImplicitParam(name = LayerVOMeta.NOTES , value = "备注" , required = false , dataTypeClass=String.class),
 	})
@@ -230,7 +276,7 @@ public class LayerController extends SuperController {
 	@ApiImplicitParams({
 		@ApiImplicitParam(name = LayerVOMeta.ID , value = "主键" , required = true , dataTypeClass=String.class , example = "565457784187916288"),
 		@ApiImplicitParam(name = LayerVOMeta.NAME , value = "名称" , required = false , dataTypeClass=String.class , example = "楼层1"),
-		@ApiImplicitParam(name = LayerVOMeta.AREA_ID , value = "区域" , required = false , dataTypeClass=String.class),
+		@ApiImplicitParam(name = LayerVOMeta.AREA_ID , value = "区域" , required = false , dataTypeClass=String.class , example = "575433967289761792"),
 		@ApiImplicitParam(name = LayerVOMeta.SORT , value = "排序" , required = false , dataTypeClass=Integer.class),
 		@ApiImplicitParam(name = LayerVOMeta.NOTES , value = "备注" , required = false , dataTypeClass=String.class),
 	})
@@ -250,64 +296,7 @@ public class LayerController extends SuperController {
 
 
 
-	/**
-	 * 导出 Excel
-	 * */
-	@SentinelResource(value = LayerServiceProxy.EXPORT_EXCEL , blockHandlerClass = { SentinelExceptionUtil.class } , blockHandler = SentinelExceptionUtil.HANDLER )
-	@RequestMapping(LayerServiceProxy.EXPORT_EXCEL)
-	public void exportExcel(LayerVO  sample,HttpServletResponse response) throws Exception {
-		try{
-			//生成 Excel 数据
-			ExcelWriter ew=layerService.exportExcel(sample);
-			//下载
-			DownloadUtil.writeToOutput(response,ew.getWorkBook(),ew.getWorkBookName());
-		} catch (Exception e) {
-			DownloadUtil.writeDownloadError(response,e);
-		}
-	}
 
-
-	/**
-	 * 导出 Excel 模板
-	 * */
-	@SentinelResource(value = LayerServiceProxy.EXPORT_EXCEL_TEMPLATE , blockHandlerClass = { SentinelExceptionUtil.class } , blockHandler = SentinelExceptionUtil.HANDLER )
-	@RequestMapping(LayerServiceProxy.EXPORT_EXCEL_TEMPLATE)
-	public void exportExcelTemplate(HttpServletResponse response) throws Exception {
-		try{
-			//生成 Excel 模版
-			ExcelWriter ew=layerService.exportExcelTemplate();
-			//下载
-			DownloadUtil.writeToOutput(response, ew.getWorkBook(), ew.getWorkBookName());
-		} catch (Exception e) {
-			DownloadUtil.writeDownloadError(response,e);
-		}
-	}
-
-
-
-	@SentinelResource(value = LayerServiceProxy.IMPORT_EXCEL , blockHandlerClass = { SentinelExceptionUtil.class } , blockHandler = SentinelExceptionUtil.HANDLER )
-	@RequestMapping(LayerServiceProxy.IMPORT_EXCEL)
-	public Result importExcel(MultipartHttpServletRequest request, HttpServletResponse response) throws Exception {
-
-		//获得上传的文件
-		Map<String, MultipartFile> map = request.getFileMap();
-		InputStream input=null;
-		for (MultipartFile mf : map.values()) {
-			input=StreamUtil.bytes2input(mf.getBytes());
-			break;
-		}
-
-		if(input==null) {
-			return ErrorDesc.failure().message("缺少上传的文件");
-		}
-
-		List<ValidateResult> errors=layerService.importExcel(input,0,true);
-		if(errors==null || errors.isEmpty()) {
-			return ErrorDesc.success();
-		} else {
-			return ErrorDesc.failure().message("导入失败").data(errors);
-		}
-	}
 
 
 }
